squid_precacher/requirements.txt
flask
flask-cors
openai
googlesearch-python
requests
configparser